requirements.txt
flask==3.0.3
tensorflow==2.17.0
pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
werkzeug==3.0.3
